Find neighbours based on Cosine similarity

+++ src/main/java/nl/lijstr/api/movies/models/MovieShortRating.java
@@ -11,7 +11,7 @@ import nl.lijstr.domain.movies.MovieRating;
 public class MovieShortRating {
 
     private int seen;
-    private BigDecimal rating;
+    private Double rating;
     private String comment;
     private long user;
