+++ src/main/java/nl/lijstr/api/movies/models/post/MovieRatingRequest.java
@@ -20,11 +20,11 @@ public class MovieRatingRequest {
     @Getter(AccessLevel.NONE)
     @DecimalMin(value = "1.0", message = "A rating can't be less than 1")
     @DecimalMax(value = "10.0", message = "A rating can't be more than 10")
-    private BigDecimal rating;
+    private Double rating;
 
     private String comment;
 
-    public BigDecimal getRating() {
+    public Double getRating() {
         if (seen == MovieRating.Seen.YES) {
             return rating;
         } else {
